feat: move benefits to it's own component
